Move merge-based graph functions and equal_computations
The merge-based functions were moved to `toolbox` to reduce unnecessary
cross-module dependencies (especially among core modules).  `equal_computations`
was moved from `scan_module` because it provides a basic graph object `__eq__`
implementation and it has multiple references outside of its own
module/sub-package.
