+++ tests/scan_module/test_scan_utils.py
@@ -7,20 +7,7 @@ import numpy as np
 import theano
 
 from theano import tensor
-from theano.scan_module.scan_utils import equal_computations, map_variables
-from theano.tensor.type_other import NoneConst
-
-
-def test_equal_compuations():
-    # This was a bug report by a Theano user.
-    c = NoneConst
-    assert equal_computations([c], [c])
-    m = theano.tensor.matrix()
-    max_argmax1 = theano.tensor.max_and_argmax(m)
-    max_argmax2 = theano.tensor.max_and_argmax(m)
-    assert equal_computations(max_argmax1, max_argmax2)
-
-
+from theano.scan_module.scan_utils import map_variables
 
 
 class TestMapVariables:
